dispatch: use platform specific paths for setting thread names
This adds the Windows specific path for setting the thread name.  This
may only work partially as the thread name is meant to be UCS-2, but for
now, matches the behaviour in Foundation.

Similarly, on Linux, `pthread_setname_np` takes a thread identifier to
set the name.  Add a platform specific path here as well.